🐛 fix: rename page names
